Split toolbar class into separate file
